Added EnemyTypeID's to support multiple enemy types
Since the dictionary cannot be viewed in the editor, there is a button on the GameManager attached component "Game Manager Enemy Type ID Handler" called Print, and another called Clear.
There is also a static class called "IDGenerator" which will generate unique IDs for enemyTypes based on their name.  To check the list, just use the Print button on the handler.

Added a bunch of new variables to a component called EnemyData, which will later be used to create the various enemy behaviors in the first iteration.